pyproject.toml
[project]
name = "heimdall"
version = "0.1.0"
description = "Browser automation agent that executes natural language tasks and collects context for automation script generation"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Team Heimdall" }
]
keywords = [
    "browser",
    "automation",
    "cdp",
    "chrome",
    "devtools",
    "testing",
    "agent",
    "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
dependencies = [
    "cdp-use>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
heimdall = "heimdall.cli:app"

[project.urls]
Homepage = "https://github.com/heimdall-automation/heimdall"
Documentation = "https://github.com/heimdall-automation/heimdall#readme"
Repository = "https://github.com/heimdall-automation/heimdall"
Issues = "https://github.com/heimdall-automation/heimdall/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/heimdall"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
